IPSecurity: reorganization
