--- common/ASC.IPSecurity/IPRestrictionsService.cs
@@ -1,87 +1,0 @@
-/*
- *
- * (c) Copyright Ascensio System Limited 2010-2018
- *
- * This program is freeware. You can redistribute it and/or modify it under the terms of the GNU 
- * General Public License (GPL) version 3 as published by the Free Software Foundation (https://www.gnu.org/copyleft/gpl.html). 
- * In accordance with Section 7(a) of the GNU GPL its Section 15 shall be amended to the effect that 
- * Ascensio System SIA expressly excludes the warranty of non-infringement of any third-party rights.
- *
- * THIS PROGRAM IS DISTRIBUTED WITHOUT ANY WARRANTY; WITHOUT EVEN THE IMPLIED WARRANTY OF MERCHANTABILITY OR
- * FITNESS FOR A PARTICULAR PURPOSE. For more details, see GNU GPL at https://www.gnu.org/copyleft/gpl.html
- *
- * You can contact Ascensio System SIA by email at sales@onlyoffice.com
- *
- * The interactive user interfaces in modified source and object code versions of ONLYOFFICE must display 
- * Appropriate Legal Notices, as required under Section 5 of the GNU GPL version 3.
- *
- * Pursuant to Section 7 § 3(b) of the GNU GPL you must retain the original ONLYOFFICE logo which contains 
- * relevant author attributions when distributing the software. If the display of the logo in its graphic 
- * form is not reasonably feasible for technical reasons, you must include the words "Powered by ONLYOFFICE" 
- * in every copy of the program you distribute. 
- * Pursuant to Section 7 § 3(e) we decline to grant you any rights under trademark law for use of our trademarks.
- *
-*/
-
-namespace ASC.IPSecurity;
-
-[Singletone]
-public class IPRestrictionsServiceCache
-{
-    public ICache Cache { get; set; }
-
-    private const string CacheKey = "iprestrictions";
-
-    internal readonly ICacheNotify<IPRestrictionItem> Notify;
-
-    public IPRestrictionsServiceCache(ICacheNotify<IPRestrictionItem> notify, ICache cache)
-    {
-        Cache = cache;
-        notify.Subscribe((r) => Cache.Remove(GetCacheKey(r.TenantId)), CacheNotifyAction.Any);
-        Notify = notify;
-    }
-
-    public static string GetCacheKey(int tenant)
-    {
-        return CacheKey + tenant;
-    }
-}
-
-[Scope]
-public class IPRestrictionsService
-{
-    private readonly ICache _cache;
-    private readonly ICacheNotify<IPRestrictionItem> _notify;
-    private readonly IPRestrictionsRepository _ipRestrictionsRepository;
-    private static readonly TimeSpan _timeout = TimeSpan.FromMinutes(5);
-
-    public IPRestrictionsService(
-        IPRestrictionsRepository iPRestrictionsRepository,
-        IPRestrictionsServiceCache iPRestrictionsServiceCache)
-    {
-        _ipRestrictionsRepository = iPRestrictionsRepository;
-        _cache = iPRestrictionsServiceCache.Cache;
-        _notify = iPRestrictionsServiceCache.Notify;
-    }
-
-    public IEnumerable<IPRestriction> Get(int tenant)
-    {
-        var key = IPRestrictionsServiceCache.GetCacheKey(tenant);
-        var restrictions = _cache.Get<List<IPRestriction>>(key);
-        if (restrictions == null)
-        {
-            restrictions = _ipRestrictionsRepository.Get(tenant);
-            _cache.Insert(key, restrictions, _timeout);
-        }
-
-        return restrictions;
-    }
-
-    public IEnumerable<string> Save(IEnumerable<string> ips, int tenant)
-    {
-        var restrictions = _ipRestrictionsRepository.Save(ips, tenant);
-        _notify.Publish(new IPRestrictionItem { TenantId = tenant }, CacheNotifyAction.InsertOrUpdate);
-
-        return restrictions;
-    }
-}
